feat: move playwright config at check level
